+++ iris-mpc/src/batch.rs
@@ -2,7 +2,6 @@ use crate::{
     aws::sns,
     iris_processing::{get_iris_shares_parse_task, IrisShareProcessor},
     message_receiver::MessageReceiver,
-    utils::handle_circuit_breaker_message,
 };
 use aws_sdk_s3::Client as S3Client;
 use aws_sdk_sns::{types::MessageAttributeValue, Client as SNSClient};
@@ -14,9 +13,9 @@ use iris_mpc_common::{
         key_pair::SharesEncryptionKeyPairs,
         shutdown_handler::ShutdownHandler,
         smpc_request::{
-            IdentityDeletionRequest, ReAuthRequest, ReceiveRequestError, SQSMessage,
-            UniquenessRequest, CIRCUIT_BREAKER_MESSAGE_TYPE, IDENTITY_DELETION_MESSAGE_TYPE,
-            REAUTH_MESSAGE_TYPE, UNIQUENESS_MESSAGE_TYPE,
+            CircuitBreakerRequest, IdentityDeletionRequest, ReAuthRequest, ReceiveRequestError,
+            SQSMessage, UniquenessRequest, CIRCUIT_BREAKER_MESSAGE_TYPE,
+            IDENTITY_DELETION_MESSAGE_TYPE, REAUTH_MESSAGE_TYPE, UNIQUENESS_MESSAGE_TYPE,
         },
         smpc_response::{
             ReAuthResult, UniquenessResult, ERROR_FAILED_TO_PROCESS_IRIS_SHARES,
@@ -475,4 +474,36 @@ pub async fn receive_batch(
     );
 
     Ok(Some(batch_builder.build()))
+}
+
+pub async fn handle_circuit_breaker_message(
+    message: &SQSMessage,
+    sqs_message: &aws_sdk_sqs::types::Message,
+    client: &Client,
+    queue_url: &str,
+    batch_size: &mut usize,
+    max_batch_size: usize,
+) -> eyre::Result<(), ReceiveRequestError> {
+    let circuit_breaker_request: CircuitBreakerRequest = serde_json::from_str(&message.message)
+        .map_err(|e| ReceiveRequestError::json_parse_error("circuit_breaker_request", e))?;
+
+    metrics::counter!("request.received", "type" => "circuit_breaker").increment(1);
+
+    client
+        .delete_message()
+        .queue_url(queue_url)
+        .receipt_handle(sqs_message.receipt_handle.clone().unwrap())
+        .send()
+        .await
+        .map_err(ReceiveRequestError::FailedToDeleteFromSQS)?;
+
+    if let Some(new_batch_size) = circuit_breaker_request.batch_size {
+        *batch_size = new_batch_size.clamp(1, max_batch_size);
+        tracing::info!(
+            "Updating batch size to {} due to circuit breaker message",
+            new_batch_size
+        );
+    }
+
+    Ok(())
 }
